webkit-efl/LayoutTests/platform/efl-tizen/fast/events/scale-and-scroll-body-expected.txt

FAIL window.document.body.scrollTop should be 30. Was 0.
FAIL window.document.body.scrollLeft should be 30. Was 0.
FAIL window.document.body.scrollTop should be 100. Was 0.
FAIL window.document.body.scrollLeft should be 100. Was 0.
FAIL window.scrollX should be 100. Was 0.
FAIL window.scrollY should be 100. Was 0.
